adapt hiplayer to media_standard.

+++ engine/pipeline/filters/demux/demuxer_filter.cpp
@@ -219,7 +219,7 @@ bool DemuxerFilter::Configure(const std::string& inPort, const std::shared_ptr<c
     return upstreamMeta->GetString(Plugin::MetaID::MEDIA_FILE_EXTENSION, uriSuffix_);
 }
 
-ErrorCode DemuxerFilter::SeekTo(int64_t msec)
+ErrorCode DemuxerFilter::SeekTo(int32_t msec)
 {
     if (!plugin_) {
         MEDIA_LOG_E("SeekTo failed due to no valid plugin");
